server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
redis==5.0.1
slowapi==0.1.9
numpy==1.24.3
cryptography==41.0.7
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1